+++ DirectIO/read_dio.cpp
@@ -17,9 +17,9 @@ int ReadDirect::rdopenfd(const fs::path &fn, int readflags) {
     return fd;
 }
 
-void ReadDirect::BlockingDirectReadAligned(const fs::path &fn, char *x, size_t length, off_t fileoffsetbytes) {
+void ReadDirect::BlockingDirectReadAligned(const fs::path &fn, char *x, size_t length, size_t fileoffsetbytes) {
     size_t lengthblocks = length/4096;          assert( lengthblocks * 4096 == length );
-    off_t offsetblocks = fileoffsetbytes/4096; assert( offsetblocks * 4096 == fileoffsetbytes );
+    size_t offsetblocks = fileoffsetbytes/4096; assert( offsetblocks * 4096 == fileoffsetbytes );
 
     int fd = rdopenfd(fn, rdopenflags() );
 
@@ -53,8 +53,8 @@ void ReadDirect::BlockingDirectReadAligned(const fs::path &fn, char *x, size_t l
     close(fd);
 }
 
-void ReadDirect::Blockingfread(const fs::path &fn, char *x, size_t length, off_t fileoffsetbytes) {
-    if (fileoffsetbytes + (off_t) length > fs::file_size(fn)){
+void ReadDirect::Blockingfread(const fs::path &fn, char *x, size_t length, size_t fileoffsetbytes) {
+    if (fileoffsetbytes + length > fs::file_size(fn)){
         fmt::print(stderr, "Trying to read offset {:d} of length {:d} from file {} of size {:d}\n",(intmax_t) fileoffsetbytes, (intmax_t) length, fn, (intmax_t) fs::file_size(fn));
         assert(0);
     }
@@ -62,7 +62,7 @@ void ReadDirect::Blockingfread(const fs::path &fn, char *x, size_t length, off_t
     FILE *fp = fopen(fn.c_str(),"rb");
     assert(fp!=NULL);
     errno = 0;
-    fseek(fp,fileoffsetbytes, SEEK_SET);
+    fseek(fp, (off_t) fileoffsetbytes, SEEK_SET);
     FAILERRNO;
     errno = 0;
     size_t bytesread = fread( x ,  sizeof(char), length, fp);
@@ -76,8 +76,8 @@ void ReadDirect::Blockingfread(const fs::path &fn, char *x, size_t length, off_t
     fclose(fp);
 }
 
-void ReadDirect::BlockingReadDirect(const fs::path &fn, char *x, size_t length, off_t fileoffsetbytes) {
-    assert( fileoffsetbytes + (off_t) length <= fs::file_size(fn) ); // can't read off the end of the file
+void ReadDirect::BlockingReadDirect(const fs::path &fn, char *x, size_t length, size_t fileoffsetbytes) {
+    assert( fileoffsetbytes + length <= fs::file_size(fn) ); // can't read off the end of the file
 
     size_t bytesleft = length;
     size_t memoryoffsetbytes = 0;
@@ -112,11 +112,11 @@ void ReadDirect::BlockingReadDirect(const fs::path &fn, char *x, size_t length, 
     Blockingfread( fn, &(x[memoryoffsetbytes]), bytesleft, fob );
 }
 
-void ReadDirect::BlockingRead(const fs::path &fn, char *x, size_t length, off_t fileoffsetbytes) {
+void ReadDirect::BlockingRead(const fs::path &fn, char *x, size_t length, size_t fileoffsetbytes) {
     ReadDirect::BlockingRead(fn, x, length, fileoffsetbytes, !allow_directio);
 }
 
-void ReadDirect::BlockingRead(const fs::path &fn, char *x, size_t length, off_t fileoffsetbytes, int no_dio) {
+void ReadDirect::BlockingRead(const fs::path &fn, char *x, size_t length, size_t fileoffsetbytes, int no_dio) {
     if(no_dio || !allow_directio) 
         Blockingfread(fn,x,length,fileoffsetbytes);
     else
